requirements.txt
Flask
Flask-SQLAlchemy
pymysql
python-dotenv
sqlalchemy
pandas
sshtunnel
ipykernel
matplotlib
seaborn
prophet
scikit-learn
statsmodels
plotly
pmdarima
requests
openai
langchain
langgraph
langsmith
langchain-anthropic
langchain-community